requirements.txt
crewai
dotenv-python
rq
sqlalchemy
psycopg2-binary
